+++ db/migrate/20170708230736_add_speed_to_rental.rb
@@ -1,9 +1,9 @@
 class AddSpeedToRental < ActiveRecord::Migration
   def up
-    add_column :rentals, :speed, :integer
+    add_column :rentals, :speed, :float
   end
 
   def down
-    remove_column :rentals, :speed, :integer
+    remove_column :rentals, :speed, :float
   end
 end
